--- a/Gemfile
+++ b/Gemfile
@@ -86,4 +86,5 @@ end
 
 group :test do
   gem 'shoulda-matchers'
+  gem 'test-prof'
 end