--- a/spec/requests/api/v1/users_controller_spec.rb
+++ b/spec/requests/api/v1/users_controller_spec.rb
@@ -4,23 +4,19 @@ require 'rails_helper'
 
 RSpec.describe API::V1::UsersController do
   describe 'GET /current' do
-    let(:params) { {} }
-    let(:headers) { {} }
-    let(:user) { create(:user) }
+    let_it_be(:user) { create(:user) }
 
     include_context 'with a valid access token', :get, :api_v1_current_user_path
 
     it_behaves_like 'an endpoint that requires token authentication'
 
-    it 'renders a successful response' do
-      call_endpoint
-      expect(response).to be_successful
-    end
-
-    it 'renders the serialized data' do
+    it 'renders a successful response and the serialized data' do
       allow(UserBlueprint).to receive(:render)
       call_endpoint
-      expect(UserBlueprint).to have_received(:render).with(user)
+      aggregate_failures do
+        expect(response).to be_successful
+        expect(UserBlueprint).to have_received(:render).with(user)
+      end
     end
   end
 end