… .rubocop.yml
… Gemfile.lock
… README.md
… spec/requests/oauth/authorization_grants_controller_spec.rb
… spec/requests/oauth/authorizations_controller_spec.rb
… spec/requests/oauth/sessions_controller_spec.rb
… spec/services/state_token_encoder_service_spec.rb
… spec/support/shared/claim_validatable.rb
… spec/support/shared/client_authentication.rb
… spec/support/shared/token_authentication.rb
… spec/support/shared/user_authentication.rb
… spec/views/oauth/authorization_grants/new.html.tailwindcss_spec.rb
… spec/views/oauth/client_error.html.tailwindcss_spec.rb