## Description

This PR adds the `test-prof` gem for evaluating test suite performance and updates the README to include its instructions. It also refactors each spec to demonstrate improvements one can make using the tool.

Previous execution time:

```
Finished in 1 minute 27.56 seconds (files took 0.81998 seconds to load)
```

Execution time after only refactoring the worst performing spec (`spec/models/oauth_session_spec.rb`):

```
Finished in 51.87 seconds (files took 0.887 seconds to load)
```

Execution time after all changes on this PR:

```
Finished in 19.21 seconds (files took 0.91638 seconds to load)
```

## Testing

Follow these steps to test this PR:

* Run the test suite on master and record the total execution time.
* Run the test suite on this branch and record the total execution time.
* Experiment with some of the commands listed in the README.